% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSRICalc-package.R
\docType{package}
\name{PSRICalc}
\alias{PSRICalc-package}
\alias{PSRICalc}
\title{PSRICalc: Plant Stress Response Index Calculator}
\description{
Calculate Plant Stress Response Index (PSRI) from germination data
}
\details{
The package provides clean, direct PSRI calculation methodology
}
\author{
\strong{Maintainer}: Richard Feiss \email{feiss026@umn.edu}

Other contributors:
\itemize{
  \item University of Minnesota [copyright holder]
}

}
